+++ src/app/controllers/account.js
@@ -1,4 +1,4 @@
-const db = require("../utils/db");
+const db = require("../../utils/db");
 const Account = db.account;
 const Op = db.Sequelize.Op;
 const jwt = require("jsonwebtoken");
