+++ src/app/models/model_room_film.js
